% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arctan_true_paraNLME.R
\name{arctan.results}
\alias{arctan.results}
\title{Title}
\usage{
arctan.results(nsim = 100, n = 80, res = results, num.interp = 45,
  true.gam1 = (2.45/pi), true.gam3 = (pi/1.1), true.gam4 = 0.8,
  beta = 0.1, beta0 = 2, file = "nlme_arctan_")
}
\arguments{
\item{file}{}
}
\value{
summarize simulation results from the parametric NLME estimation, where the true data is generated by the
        arctangent model and the estimation is performed under the artangent model.
}
\description{
Title
}
\examples{
 simus.arctan.out<-arctan.results(nsim=100, n=80, res=results, num.interp=45, true.gam1=(2.45/pi), true.gam3=(pi/1.1), true.gam4=0.8,  beta=0.1, beta0=2,  file="nlme_arctan_")



}
