% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actan_nlme.R
\name{arctan_first_deriv_ft}
\alias{arctan_first_deriv_ft}
\title{Title}
\usage{
arctan_first_deriv_ft(gam1, gam2, gam3, logage)
}
\arguments{
\item{logage}{}
}
\value{
a vector of the first derivatives of the arctangent function.
}
\description{
Title
}
\examples{
 logage<-seq(-0.5, 0.5, by=0.1);
           arctan_first_derv<-arctan_first_deriv_ft(2.45/pi, 0, pi/1.1, 0.8, logage);
           plot(logage, arctan_first_derv, type='l')

}
