% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actan_nlme.R
\name{arctan_second_deriv_ft}
\alias{arctan_second_deriv_ft}
\title{Second derivative of the arctangent model function for the parametric nonlinear mixed effects model (nlme) procedure}
\usage{
arctan_second_deriv_ft(gam1, gam2, gam3, logage)
}
\arguments{
\item{gam1}{a scale parameter.}

\item{gam2}{a parameter for an inflection point.}

\item{gam3}{a parameter, which determines steepness of the arctangent model.}

\item{logage}{a vector of time points (log-scaled ages), which is generated around \code{gam2}. It can be an any length of numeric vector.}
}
\value{
a vector of the second derivative of the arctangent function with respect to (logage-gam2).
}
\description{
Second derivative of the arctangent model function for the parametric nonlinear mixed effects model (nlme) procedure
}
\examples{

library(HDChangePoint)

## Specify parameters
gam1=2.45/pi
gam2=0
gam3=pi/1.1

## Specify the time points (log-scaled ages)
logage<-seq(-0.5, 0.5, by=0.1)

## Obtain a vector of the second derivative of the arctangent model
arct_sec_derv<-arctan_second_deriv_ft(2.45/pi, 0, pi/1.1, logage)

## Plot the second derivative of the arctangent model
plot(logage, arct_sec_derv, type='l')



}
