% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actan_nlme.R
\name{arctanf}
\alias{arctanf}
\title{Title}
\usage{
arctanf(gam1, gam2, gam3, gam4, logage)
}
\arguments{
\item{logage}{}
}
\value{
a vector of variables from the arctangent function.
}
\description{
Title
}
\examples{
 logage<-seq(-0.5, 0.5, by=0.1);
           arctanft<-arctanf(2.45/pi, 0, pi/1.1, 0.8, logage);
           plot(logage, arctanft, type='l')


}
