% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpara_derivative.R
\name{finite.differences}
\alias{finite.differences}
\title{Title}
\usage{
finite.differences(x, y)
}
\arguments{
\item{y}{}
}
\value{
Given vectors x and y, returns first derivatives of vector y corresponding a vector x
}
\description{
Title
}
\examples{
          x<-seq(-0.5, 0.5, by=0.1); y<-w(x, 0, model="logist");
          out<-finite.differences(x, y); plot(out);



}
