% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_analysis.R
\name{hd.study}
\alias{hd.study}
\title{Analysis example using the simulated HD dataset}
\usage{
hd.study(simu.data = simu.data, m = 30, num.interp = 30, n = 80,
  newl = 30, mean.diff = 1, tolerance = 0.005, itermax = 20,
  iter = 0)
}
\arguments{
\item{m}{number of time points at which predictors are required for the longitudinal responses in the parametric NLME procedure}

\item{num.interp}{number of pseudo-data points, which are unifromly generated within the observed data points.}

\item{n}{number of sample size.}

\item{newl}{a length of time points (log transformed ages) at which predictors are required for each individual longitudinal trajectory.}

\item{mean.diff}{an initial tolerence for the convergence criterion. Default is 1.}

\item{tolerance}{a parameter for the tolerence for the convergence criterion. Default is 0.009.}

\item{itermax}{a parameter for the maximum number of iterations in the multi-stage nonparametric algorithm. Default is 20.}

\item{iter}{initial number of iteration in the multi-stage nonparametric algorithm. The default is 0.}

\item{simu.dat}{a simulated data set, see \code{DATA.R} for details.}
}
\value{
a list of
       \item{nonpara_summary_table4}{a 5 x 4 array of estimates, standard errors, t-values and p-values for the fixed effects of \code{beta0}, CAG repeates and gender in the inflection point model and scale parameters of the random errors in the inflection point and the longitudinal models.}
       \item{para_summary_table4}{a 7 x 4 array of estimates, standard errors, t-values and p-values for the fixed effects of \code{theta1}, \code{theta2}, \code{beta0}, CAG repeates and gender and scale parameters of the random errors in the inflection point and the longitudinal models. The parametric NLME assumes the logistic model.}
       , which are similar to Table 4 and return Figure 1 in the paper (See the \code{reference}).
}
\description{
Analysis example using the simulated HD dataset
}
\examples{

library(HDChangePoint)

## a pseudo data set constructed by using the simulated data
data("PSEUDO_PREDICT_HD")
head(PSEUDO_PREDICT_HD)

## Specify the parameters to obtained the analysis results from the simulated dataset.
simu.dat<-PSEUDO_PREDICT_HD
n=80;
m=45;
num.interp=45;
newl=45;
mean.diff=1;
tolerance=0.01;
itermax=20;
iter=0;

simu.analysis.results<-hd.study(simu.data=simu.data, m=m, num.interp=num.interp, n=n, newl=newl,
                               mean.diff=mean.diff, tolerance=tolerance, itermax=itermax, iter=iter)

Also, see man/examples/reproduce_simulation_study_analysis.R


}
\references{
U.Lee, R.J.Carroll, K.Marder, Y.Wang, T.P.Garcia. (2018+). Estimating Disease Onset from Change Points of Markers Measured with Error.
}
