% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_nlme.R
\name{logist_second_deriv_ft}
\alias{logist_second_deriv_ft}
\title{Title}
\usage{
logist_second_deriv_ft(theta1, theta2, theta3, logage)
}
\arguments{
\item{logage}{}
}
\value{
a vector of the second derivatives of the logistic function
}
\description{
Title
}
\examples{
 logage<-seq(-0.5, 0.5, by=0.1);
           logit_sec_derv<-logist_second_deriv_ft(2.45/pi, 0, pi/1.1, 0.8, logage);
           plot(logage, logit_sec_derv, type='l')



}
