% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_nlme.R
\name{logistft}
\alias{logistft}
\title{Title}
\usage{
logistft(theta1, theta2, theta3, logage)
}
\arguments{
\item{logage}{}
}
\value{
a vector of variables from the logistic function.
}
\description{
Title
}
\examples{
 logage<-seq(-0.5, 0.5, by=0.1);
           logit.ft<-logistft(6, 0, 1, logage);
           plot(logage, logit.ft, type='l')

}
