% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_true_paraNLME.R
\name{logit.results}
\alias{logit.results}
\title{Title}
\usage{
logit.results(nsim = 100, n = 80, res = results, theta1 = 6,
  theta3 = 1, beta = 0.1, beta0 = 0.5, num.interp = 45,
  file = "nlme_logist")
}
\arguments{
\item{file}{}
}
\value{
summarize simulation results from the parametric NLME estimation, where the true data is generated by the
        logistic model and the estimation is performed under the logistic model.
}
\description{
Title
}
\examples{
 simus.logist.out<-logit.results(nsim=100, n=80, res=results, theta1=6, theta3=1, beta=0.1, beta0=0.5, num.interp=45, file="nlme_logist")



}
