% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arctan_missp_paraNLME.R
\name{main.arctan.missp.nlme}
\alias{main.arctan.missp.nlme}
\title{Title}
\usage{
main.arctan.missp.nlme(n = 80, model = "logist", dat = outdat,
  num.boot = 1000, num.interp = 45, true.theta1 = 6,
  true.theta3 = 1, para1 = 1, para2 = 3.5, para3 = 0, para4 = 3,
  para5 = 1, eps.sd = 0.05, dist = "normal")
}
\arguments{
\item{dist}{}
}
\value{
Generate simulation results for the parametric NLME approach, where true data are generated by the logistic model
        but the parametric NLME estimation is performed under the incorrect (arctangent) model.
}
\description{
Title
}
\examples{
 #library(nlme); library(scam);  library(ShapeChange);

            library(registry); library(pkgmaker); library(rngtools); library(iterators);
            library(foreach); library(parallel);

            libaray(doParallel); nworkers<-detectCores()
            cl<-makePSOCKcluster(nworkers);
            library(doMC); registerDoMC(nworkers)
            library(doRNG)

            set.seed(1)
            outdat<-mydata(n=80, model="logist", p=2, bb0=0.5, bb=0.1, x.sd=0.3, v1=5, v2=7, dist="normal", eps.sd=0.05, u.sd=0.05)

            #parallel computing
            results<-foreach(i=1:nsim)\%dorng\%{

                                main.arctan.missp.nlme(n=80, model="logist",  dat=outdat, num.boot=1000,num.interp=45,
                                true.theta1=6, true.theta3=1,  para1=1, para2=3.5, para3=0, para4=3, para5=1,  eps.sd=0.05, dist="normal")

                                }
            stopCluster(cl)



}
