% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_true_paraNLME.R
\name{main.logist.nlme}
\alias{main.logist.nlme}
\title{Title}
\usage{
main.logist.nlme(n = 80, model = "logist", dat = outdat,
  num.boot = 1000, true.theta1 = 6, true.theta3 = 1,
  num.interp = 45, eps.sd = 0.05, dist = "normal", para1 = 3.5,
  para2 = 0.1, para3 = 0.1, para4 = 1)
}
\arguments{
\item{para4}{}
}
\value{
Generate simulation results for the parametric NLME approach, where true data are generated by the logistic model
        and the parametric NLME estimation is performed under the correct (logistic) model.
}
\description{
Title
}
\examples{
  #library(nlme); library(scam);  library(ShapeChange);

            library(registry); library(pkgmaker); library(rngtools); library(iterators);
            library(foreach); library(parallel);

            libaray(doParallel); nworkers<-detectCores()
            cl<-makePSOCKcluster(nworkers);
            library(doMC); registerDoMC(nworkers)
            library(doRNG)

            set.seed(1)
            outdat<-mydata(n=80, model="logist", p=2, bb0=0.5, bb=0.1, x.sd=0.3, v1=5, v2=7, dist="normal", eps.sd=0.05, u.sd=0.05)

            #parallel computing
            results<-foreach(i=1:nsim)\%dorng\%{

                                main.logist.nlme(n=80, model="logist", dat=outdat, num.boot=1000,
                                true.theta1=6, true.theta3=1, num.interp=45, eps.sd=0.05, dist="normal", para1=4.3, para2=4.1, para3=0, para4=1)

                                }
            stopCluster(cl)


}
