% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_missp_paraNLME.R
\name{main.logistic.missp.nlme}
\alias{main.logistic.missp.nlme}
\title{Title}
\usage{
main.logistic.missp.nlme(n = 80, model = "arctan", num.boot = 1000,
  dat = outdat, num.interp = 45, true.gam1 = 2.45/pi,
  true.gam3 = pi/1.1, para1 = 3.1, para2 = 0.2, para3 = 0.2,
  para4 = 1.5, eps.sd = 0.05, dist = "normal")
}
\arguments{
\item{dist}{}
}
\value{
Generate simulation results for the parametric NLME approach, where true data are generated by the arctangent model
        but the parametric NLME estimation is performed under the incorrect (logistic) model.
}
\description{
Title
}
\examples{
 #library(nlme); library(scam);  library(ShapeChange);

            library(registry); library(pkgmaker); library(rngtools); library(iterators);
            library(foreach); library(parallel);

            libaray(doParallel); nworkers<-detectCores()
            cl<-makePSOCKcluster(nworkers);
            library(doMC); registerDoMC(nworkers)
            library(doRNG)

            set.seed(1)
            outdat<-mydata(n=80, model="arctan", p=2, bb0=2, bb=0.1, x.sd=0.3, v1=5, v2=7, dist="normal", eps.sd=0.05, u.sd=0.05)

            #parallel computing
            results<-foreach(i=1:nsim)\%dorng\%{

                                main.logistic.missp.nlme(n=80, model="arctan", dat=outdat, num.boot=1000,   num.interp=45, true.gam1=2.45/pi, true.gam3=pi/1.1, para1=4.3,
                                             para2=0.1, para3=0.1, para4=1.5,  eps.sd=0.05, dist="normal")
                                }
            stopCluster(cl)


}
