% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arctan_missp_paraNLME.R
\name{missp.arctan.results}
\alias{missp.arctan.results}
\title{Title}
\usage{
missp.arctan.results(nsim = 100, n = 80, res = results, beta = 0.1,
  beta0 = 0.5, num.interp = 45, file = "nlme_arctan_missp_")
}
\arguments{
\item{file}{}
}
\value{
summarize simulation results from the parametric NLME estimation, where the true data is generated by the
        logistic model, but the estimation is performed under the arctangent model.
}
\description{
Title
}
\examples{
 simus.missp.arctan.out<-missp.arctan.results(nsim=100, n=80, res=results, num.interp=45,  beta=0.1, beta0=0.5, num.interp=45, file="nlme_arctan_")



}
