% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_missp_paraNLME.R
\name{missp.logist.results}
\alias{missp.logist.results}
\title{Title}
\usage{
missp.logist.results(nsim = 100, n = 80, res = results, beta = 0.1,
  beta0 = 2, num.interp = 45, file = "nlme_logisic")
}
\arguments{
\item{file}{}
}
\value{
summarize simulation results from the parametric NLME estimation, where the true data is generated by the
        arctangent model, but the estimation is performed under the logistic model.
}
\description{
Title
}
\examples{
 simus.missp.logist.out<-missp.logist.results(nsim=100, n=80, res=results, num.interp=45,  beta=0.1, beta0=2,  file="nlme_arctan_")


}
