% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dat.R
\name{mydata}
\alias{mydata}
\title{Title}
\usage{
mydata(n = 80, model = "logist", p = 2, bb0 = 1, bb = 1,
  x.sd = 5, v1 = 5, v2 = 7, dist = "normal", eps.sd = 0.05,
  u.sd = 0.05)
}
\arguments{
\item{u.sd}{}
}
\value{
a data frame for a simulated longitudinal data containing subject ID, subject specific covariates,
        true  inflection points, longitudinal time points, longitudinal response and gender.
}
\description{
Title
}
\examples{
mydata(n=80, model="logist", p=2, bb0=0.5, bb=0.1, x.sd=0.3, v1=5, v2=7, dist="normal", eps.sd=0.05, u.sd=0.05)



}
