% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{PSEUDO_PREDICT_HD}
\alias{PSEUDO_PREDICT_HD}
\title{Generate a simulated dataset to reproduce results in the analysis study.}
\format{A dataframe including 80 subjects' longitudinal data with 8 variables:
\describe{
  \item{n}{number of sample size}
  \item{model}{a character string to choose a nonlinear model \code{w()}, \code{"logist"} or \code{"arctan"}.}
    \item{p}{dimension of parameters corresponding to subject specific covariates including a constant term. The default is 2.}
  \item{bb0}{an intercept term for the inflection point model.}
    \item{bb}{(p-1)-length of parameters of subject specific covariates in the inflection point model.}
  \item{x.sd}{standard deviation of log transformed ages at clinical visits to generate time points.}
    \item{v1}{the least number of visits for all subjects.}
  \item{v2}{the largest number of visits for all subjects.}
   \item{dist}{distribution for the error terms for the longitudinal model.}
  \item{eps.sd}{standard deviation for the error term for the longitudinal model.}
  \item{u.sd}{standard deviation for the error term for the inflection point model.}
}}
\source{
see data-raw/PSEUDO_PREDICT_HD.R
}
\usage{
PSEUDO_PREDICT_HD
}
\description{
A pseudo dataset containing subject's id, number of clinical visits, total motor scores,
time of the clinical diagnosis, ages at clinical visits, CAG repeats, Gender, and log transformed ages at clinical visits.
}
\keyword{datasets}
