% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpara_derivative.R
\name{sec.der.mid}
\alias{sec.der.mid}
\title{Second Derivatives Using the Midpoint Formula}
\usage{
sec.der.mid(x, y)
}
\arguments{
\item{x}{a m-length of numeric vector.}

\item{y}{a m-length of numeric vector.}
}
\value{
Given m-length of vectors x and y, returns a m-length of the second derivatives of y
        corresponding a vector x.
}
\description{
Second Derivatives Using the Midpoint Formula
}
\examples{

library(HDChangePoint)

##Generate the same length of vectors x and y.

x<-seq(-0.5, 0.5, by=0.01);
y<-w(x, 0, model="logist");


## Obtain the second derivatives of y using the midpoint formula.
out<-sec.der.mid(x, y); plot(out);


}
