% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stage_nonpara.R
\name{sim.nonpara}
\alias{sim.nonpara}
\title{Title}
\usage{
sim.nonpara(n = 80, model = "arctan", num.interp = 25, newl = 25,
  dist = "normal", k1 = 10, k2 = 10, eps.sd = 0.05,
  mean.diff = 1, eps = 0.009, itermax = 20, iter = 0,
  dat = outdat)
}
\arguments{
\item{dat}{}
}
\value{
main function to generate simulation results for the multi-stage nonparametric approach including the estimated individual longitudinal
        trajectories and their standard errors, estimated inflection points, estimatimated longitudinal model parameters and their standard errors.
}
\description{
Title
}
