% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_results_nonpara.R
\name{simus.results}
\alias{simus.results}
\title{Summary results of the multi-stage nonparametric procedure}
\usage{
simus.results(nsim = 100, n = 80, res = results, model = "logist",
  num.interp = 45, newl = 45, bb.true = 0.1, b0.true = 0.5,
  true.sigma.u = 0.05, true.sigma.eps = 0.05, file = "logist")
}
\arguments{
\item{nsim}{number of simulation runs.}

\item{n}{number of sample size.}

\item{res}{a nsim-length of list from the simulated results.}

\item{model}{a character string for a nonlinear model: \code{"logist"} or \code{"arctan"}.}

\item{num.interp}{number of pseudo-data generated in the estimation procedure.}

\item{newl}{number of time points (log transformed ages) at which predictors are required for each individual longitudinal trajectory.}

\item{bb.true}{a parameter for the (p-1)-length of true coefficient vector corresponding to subject specific covariates.}

\item{b0.true}{a parameter for the true intercept of the log-normal model for the inflection point.}

\item{true.sigma.u}{a true scale parameter for the random error term in the inflection point model.}

\item{true.sigma.eps}{a true scale parameter for the within-subject error term in the longitudinal model.}

\item{file}{a character string of file name to generate plots with .eps file extention.

'}
}
\value{
A list of the summarized simulation results from the multi-stage nonparametric procedure

       \item{res.table}{a data frame of the absolute biases, estimated standard deviations, average of the estimated standard errors and 95\% coverage probabilities for the fixed effects (\code{bb.true}, \code{bb0.true}).}
       \item{avr.est.Z}{a data frame of the average performance of inflection points for all subjects including the absolute biases, the empirical standard errors, bootstrap standard deviations and 95\% boostrap confidence intervals.}
}
\description{
Summary results of the multi-stage nonparametric procedure
}
\examples{

library(HDChangePoint)


## Specify parameters to generate true data
n=80;
model="logist";
p=2;
bb0=0.5;
bb=0.1;
x.sd=0.3;
v1=5;
v2=7;
dist="normal";
eps.sd=0.05;
u.sd=0.05;

set.seed(22)
## Data generation under the logistic model
outdat<-mydata(n=n, model=model, p=p, bb0=bb0, bb=bb, x.sd=x.sd, v1=v1, v2=v2, dist=dist, eps.sd=eps.sd, u.sd=u.sd)


## Specify parameters for the multi-stage nonparametric procedure
n=80;
model="logist"
num.interp=45;
newl=45;
k1=20;
k2=20;
mean.diff=1;
tolerance=0.009
itermax=20;
iter=0;

## Multi-stage nonparametric estimation
results<-sim.nonpara(n=n, model=model, num.interp=num.interp, newl=newl,
                     dist="normal", k1=k1, k2=k2, eps.sd=eps.sd, mean.diff=mean.diff,
                     tolerance=tolerance, itermax=itermax, iter=iter, dat=outdat)


## Specify parameters to obtain the summary results when the true data were generated from the logistic model.

nsim=100;
n=80;
num.interp=45;
newl=45;
bb.true=0.1;
b0.true=0.5;
true.sigma.u=0.05;
true.sigma.eps=0.05;
file="logist";

simus.out<-simus.results(nsim=nsim, n=n, res=results, model="logist", num.interp=num.interp, newl=newl, bb.true=bb.true,
                         b0.true=b0.true, true.sigma.u=true.sigma.u, true.sigma.eps=true.sigma.eps, file=file)




}
