% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_omega_ft.R
\name{w}
\alias{w}
\title{Title}
\usage{
w(x, logT, model)
}
\arguments{
\item{model}{}
}
\value{
a vector of response values corresponding time points x, whose trajecotry follows an S-shape.
        In practice, the locations of the inflection points (logT) are latent values.
}
\description{
Title
}
\examples{
 x<-seq(-0.5, 0.5, by=0.1); logT<-0;
           out<-w(x, logT, model="logist"); plot(out)



}
