% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_results_summary_true_arctan.R
\name{arctan.results}
\alias{arctan.results}
\title{Summary results of the parametric NLME, when the true data are generated by the arctangent model and the parametric NLME assumes the correct arctangent  model.}
\usage{
arctan.results(nsim = 100, n = 80, res = results, time.length = 20,
  true.gam1 = (2.45/pi), true.gam3 = (pi/1.1), true.gam4 = 0.8,
  beta = 0.1, beta0 = 2, file = "nlme_arctan_")
}
\arguments{
\item{nsim}{number of simulation runs.}

\item{n}{number of sample size.}

\item{res}{a nsim-length of list for simulation results.}

\item{time.length}{number of data points at which predictors are required for each individual longitudinal trajectory. This time point for graphs to be plotted.}

\item{true.gam1}{a true scale parameter for the arctangent model.}

\item{true.gam3}{a true parameter, which determines the steepness of the arctangent model.}

\item{true.gam4}{a true parameter, which determines the minimum and maximum of the arctangent model.}

\item{beta}{a (p-1)-length of true coefficient vector corresponding to subject specific covariates.}

\item{beta0}{true intercept of the log-normal model for the inflection point.}

\item{file}{a character string of file name for a creating plot in .eps file extention.}
}
\value{
A list of the summarized simulation results including

       \item{res.table}{a data frame of the absolute biases, estimated standard deviations, average of the estimated standard errors and 95\% coverage probabilities for the fixed effects (\code{beta0}, \code{beta}).}
       \item{Rel.glob.logT.est}{a data frame of the relative average performance of inflection points for all subjects including the relative absolute biases, the relative empirical standard errors, the relative bootstrap standard deviations and 95\% boostrap confidence intervals.}
}
\description{
Summary results of the parametric NLME, when the true data are generated by the arctangent model and the parametric NLME assumes the correct arctangent  model.
}
\examples{

library(HDChangePoint)

data(nlme_arctan_visit1)

## Specify parameters to obtain the summary results

nsim=100;
n=80;
time.length=20;
true.gam1=(2.45/pi);
true.gam3=(pi/1.1);
true.gam4=0.8;
beta0=2;
beta=0.1;
file="nlme_arctan_";


## summary results
simus.arctan.out<-arctan.results(nsim=nsim, n=n, res=results, time.length=time.length, true.gam1=true.gam1,
                                 true.gam3=true.gam3, true.gam4=true.gam4,  beta=beta, beta0=beta0,  file=file)




}
