% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_results_summary_true_arctan.R
\name{arctan.results}
\alias{arctan.results}
\title{Summary results of the parametric NLME, when the true data are generated by the arctangent model and the parametric NLME assumes the correct arctangent  model.}
\usage{
arctan.results(nsim = 100, n = 80, res = results, num.grid = 45,
  true.gam1 = (2.45/pi), true.gam3 = (pi/1.1), true.gam4 = 0.8,
  beta = 0.1, beta0 = 2, file = "nlme_arctan_")
}
\arguments{
\item{nsim}{number of simulation runs.}

\item{n}{number of sample size.}

\item{res}{a nsim-length of list for simulation results.}

\item{num.grid}{number of data points at which predictors are required for each individual longitudinal trajectory.}

\item{true.gam1}{a true scale parameter for the arctangent model.}

\item{true.gam3}{a true parameter, which determines the steepness of the arctangent model.}

\item{true.gam4}{a true parameter, which determines the minimum and maximum of the arctangent model.}

\item{beta}{a (p-1)-length of true coefficient vector corresponding to subject specific covariates.}

\item{beta0}{true intercept of the log-normal model for the inflection point.}

\item{file}{a character string of file name for a creating plot in .eps file extention.}
}
\value{
A list of the summarized simulation results including

       \item{res.table}{a data frame of the absolute biases, estimated standard deviations, average of the estimated standard errors and 95\% coverage probabilities for the fixed effects (\code{beta0}, \code{beta}).}
       \item{avr.est.Z}{a data frame of the average performance of inflection points for all subjects including the absolute biases, the empirical standard errors, bootstrap standard deviations and 95\% boostrap confidence intervals.}
}
\description{
Summary results of the parametric NLME, when the true data are generated by the arctangent model and the parametric NLME assumes the correct arctangent  model.
}
\examples{

library(HDChangePoint)

## Specify parameters to generate true data
n=80;
model="arctan";
p=2;
bb0=2;
bb=0.1;
x.sd=0.3;
v1=5;
v2=7;
dist="normal";
eps.sd=0.05;
u.sd=0.05;

## generate data with seed number
set.seed(1)
outdat<-mydata(n=n, model=model, p=p, bb0=bb0, bb=bb, x.sd=x.sd, v1=v1, v2=v2, dist=dist, eps.sd=eps.sd, u.sd=u.sd)



## Specify parameters for the parametric NLME procedure

num.boot=1000;
true.gam1=2.45/pi;
true.gam3=pi/1.1;
num.grid=45;
eps.sd=0.05;
dist="normal";
para1=0.8;
para2=6.1;
para3=0;
para4=2.8;
para5=1.5;


## Do parametric NLME estimation
results<-main.arctan.nlme(n=n, model=model, dat=outdat, num.boot=num.boot, num.grid=num.grid, true.gam1=true.gam1,
                          true.gam3=true.gam3, para1=para1, para2=para2, para3=para3, para4=para4, para5=para5,
                          eps.sd=eps.sd, dist=dist)


## Specify parameters to obtain the summary results

nsim=100;
n=80;
num.grid=45;
true.gam1=(2.45/pi);
true.gam3=(pi/1.1);
true.gam4=0.8;
beta0=2;
beta=0.1;
file="nlme_arctan_";


## summary results
simus.arctan.out<-arctan.results(nsim=nsim, n=n, res=results, num.grid=num.grid, true.gam1=true.gam1,
                                 true.gam3=true.gam3, true.gam4=true.gam4,  beta=beta, beta0=beta0,  file=file)



}
