% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actan_nlme.R
\name{arctan_second_deriv_ft}
\alias{arctan_second_deriv_ft}
\title{Title}
\usage{
arctan_second_deriv_ft(gam1, gam2, gam3, logage)
}
\arguments{
\item{logage}{}
}
\value{
a vector of the second derivatives of the arctangent function
}
\description{
Title
}
\examples{
 logage<-seq(-0.5, 0.5, by=0.1);
           arct_sec_derv<-arctan_second_deriv_ft(2.45/pi, 0, pi/1.1, 0.8, logage);
           plot(logage, arct_sec_derv, type='l')

}
