% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpara_derivative.R
\name{finite.differences}
\alias{finite.differences}
\title{First Derivative Using A Finite Differences Method}
\usage{
finite.differences(x, y)
}
\arguments{
\item{x}{a m-length of numeric vector.}

\item{y}{a m-length of numeric vector.}
}
\value{
Given vectors x and y, returns a m-length of the first derivatives of vector y corresponding a vector x. The reference is \url{https://www.r-bloggers.com/numerical-differentiation-with-finite-differences-in-r/}.
}
\description{
First Derivative Using A Finite Differences Method
}
\examples{


library(HDChangePoint)

##Generate the same length of vectors x and y.
x<-seq(-0.5, 0.5, by=0.01);
y<-w(x, 0, model="logist");

## Obtain the first derivatives of y using the forward difference method.
out<-finite.differences(x, y); plot(out);



}
