% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_nlme.R
\name{logistft}
\alias{logistft}
\title{Logistic model for the parametric nonlinear mixed effects model (nlme) procedure}
\usage{
logistft(theta1, theta2, theta3, logage)
}
\arguments{
\item{theta1}{a scale parameter, which determines steepness of the logistic model.}

\item{theta2}{a parameter for an inflection point.}

\item{theta3}{a parameter, which determines the maximum  (asymptote) of the logistic model.}

\item{logage}{a vector of time points (log-scaled ages), which is generated around \code{theta2}. It can be an any length of numeric vector.}
}
\value{
a vector of response variables under the deterministic logistic model, whose length is the same as the length of \code{logage}.
}
\description{
Logistic model for the parametric nonlinear mixed effects model (nlme) procedure
}
\examples{

library(HDChangePoint)

## Specify parameters
theta1=6;
theta2=0;
theta3=1;


## Specify the time points (log-scaled ages)
logage<-seq(-0.5, 0.5, by=0.1)

## Obtain a vector of response variables when parametric NLME assumes the logistic model
logit.ft<-logistft(6, 0, 1, logage)

## Plot the S-shaped curve under the logistic model
plot(logage, logit.ft, type='l')


}
