% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_results_summary_true_logist.R
\name{logit.results}
\alias{logit.results}
\title{Summary results of the parametric NLME, when the true data are generated by the logistic model and the parametric NLME assumes the correct logistic  model.}
\usage{
logit.results(nsim = 100, n = 80, res = results, theta1 = 6,
  theta3 = 1, beta = 0.1, beta0 = 0.5, num.grid = 45,
  file = "nlme_logist")
}
\arguments{
\item{nsim}{number of simulation runs.}

\item{n}{number of sample size.}

\item{res}{a nsim-length of list for simulation results.}

\item{theta1}{a true scale parameter, which determines steepness of the logistic model.}

\item{theta3}{a true parameter, which determines the maximum (asymptote) of logstic model.}

\item{beta}{a (p-1)-length of true coefficient vector corresponding to subject specific covariates.}

\item{beta0}{true intercept of the log-normal model for the inflection point.}

\item{num.grid}{number of data points at which predictors are required for each individual longitudinal trajectory.}

\item{file}{a character string of file name for a creating plot in .eps file extention.}
}
\value{
A list of the summarized simulation results including

       \item{res.table}{a data frame of the absolute biases, estimated standard deviations, average of the estimated standard errors and 95\% coverage probabilities for the fixed effects (\code{beta0}, \code{beta}).}
       \item{avr.est.Z}{a data frame of the average performance of inflection points for all subjects including the absolute biases, the empirical standard errors, bootstrap standard deviations and 95\% boostrap confidence intervals.}
}
\description{
Summary results of the parametric NLME, when the true data are generated by the logistic model and the parametric NLME assumes the correct logistic  model.
}
\examples{


library(HDChangePoint)


## Specify parameters to generate true data
n=80;
model="logist";
p=2;
bb0=0.5;
bb=0.1;
x.sd=0.3;
v1=5;
v2=7;
dist="normal";
eps.sd=0.05;
u.sd=0.05;

## generate data with seed number
set.seed(1)
outdat<-mydata(n=n, model=model, p=p, bb0=bb0, bb=bb, x.sd=x.sd, v1=v1, v2=v2, dist=dist, eps.sd=eps.sd, u.sd=u.sd)



## Specify parameters for the parametric NLME procedure
n=80;
model="logist";
num.boot=1000;
true.theta1=6;
true.theta3=1;
num.grid=45;
eps.sd=0.05;
dist="normal";
para1=4.3;
para2=4.1;
para3=0;
para4=1;

## Do parametric NLME estimation
results<-main.logist.nlme(n=n, model=model, dat=outdat, num.boot=num.boot,
                          true.theta1=true.theta1, true.theta3=true.theta3, num.grid=num.grid,
                          eps.sd=eps.sd, dist=dist, para1=para1, para2=para2, para3=para3, para4=para4)




## Specify parameters to obtain the summary results

nsim=100;
n=80;
theta1=6;
theta3=1;
beta0=0.5;
beta=0.1;
num.grid=45;
file="nlme_logist";

## summary results
simus.logist.out<-logit.results(nsim=nsim, n=n, res=results, theta1=theta1, theta3=theta3, beta=beta, beta0=beta0,
                                num.grid=num.grid, file=file)



}
