% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arctan_missp_paraNLME.R
\name{main.arctan.missp.nlme}
\alias{main.arctan.missp.nlme}
\title{Parametric nonlinear mixed effects model (NLME) approach: When true data are generated by the logistic model, the parametric NLME procedure is performed under the (missepcified) arctangent model .}
\usage{
main.arctan.missp.nlme(n = 80, model = "logist", dat = outdat,
  num.boot = 1000, num.grid = 45, true.theta1 = 6, true.theta3 = 1,
  para1 = 1, para2 = 3.5, para3 = 0, para4 = 3, para5 = 1,
  eps.sd = 0.05, dist = "normal")
}
\arguments{
\item{n}{number of sample size.}

\item{model}{a character string for a nonlinear model: \code{"logist"} or \code{"arctan"}.}

\item{dat}{a data frame of the generated data set.}

\item{num.boot}{number of bootsrap replicates.}

\item{num.grid}{number of data points at which predictors are required for each individual longitudinal trajectory.}

\item{true.theta1}{a true parameter for \code{theta1}, which determines steepness of the logistic function, see \code{logistft()}.}

\item{true.theta3}{a true parameter for \code{theta3}, which determines the maximum  (asymptote) of the logistic function, see \code{logistft()}.}

\item{para1}{an initial parameter for \code{gam1}, which is used for the parametric nonlinear mixed effects model (NLME) approach.}

\item{para2}{an initial intercept parameter for the inflection point \code{gam2}, which is used for the parametric nonlinear mixed effects model (NLME) approach.}

\item{para3}{an initial (p-1)-length of coefficient vector of subject specific covariates for the inflection point \code{gam2}, which is used for the parametric nonlinear mixed effects model (NLME) approach.}

\item{para4}{an initial parameter for \code{gam3}, which is used for the parametric nonlinear mixed effects model (NLME) approach.}

\item{para5}{an initial vertical shift parameter for \code{gam4}, which is used for the parametric nonlinear mixed effects model (NLME) approach.}

\item{eps.sd}{a true scale parameter of the within-subject error term in the longitudinal model.}

\item{dist}{a character string for the distribution of within-subject error term in the longitudinal model. Default is \code{"normal"}.}
}
\value{
A list of

       \item{est.beta}{the (p-1)-length of estimated coefficient vector of subjec-specific covariates in the log-normal model for inflection points.}
       \item{est.beta0}{estimated intercept of the log-normal model for inflection points.}
       \item{est.str.beta}{the (p-1)-length of estimated standard errors of the coefficient vector of subject-specific covariates in the log-normal model for inflection points.}
       \item{est.str.beta0}{estimated standard errors of the intercept of the log-normal model for inflection points.}
       \item{cp.lower.beta0}{estimated lower bound of the 95\% confidence interval for beta0, which is the intercept of the log-normal model for inflection points.}
       \item{cp.upper.beta0}{estimated upper bound of the 95\% confidence interval for beta0, which is the intercept of the log-normal model for inflection points.}
       \item{cp.lower.beta}{estimated lower bound of the 95\% confidence interval for beta, which is the (p-1)-length of the coefficient vector of subject specific covariates in the log-normal model for inflection points.}
       \item{cp.upper.beta}{estimated upper bound of the 95\% confidence interval for beta, which is the (p-1)-length of the coefficient vector of subject specific covariates in the log-normal model for inflection points.}
       \item{sig.eps}{estimated scale parameter of the within-subject error in the longitudinal model.}
       \item{sig.u}{estimated scale parameter of random error in the log normal model for the inflection points.}
       \item{est.rand.ef}{estimated random effects in the log normal model for the inflection points.}
       \item{est.logT}{the n-length of the estimated inflection points vector, where each element is the individual estimated inflection point.}
       \item{true.logT}{the n-length of the true inflection points vector, where each element is the true individual inflection point.}
       \item{new.pred.data}{a (num.grid/2) x 5 x n array of data set to generate boostrap estimates, which includes ID, log scaled ages, subject specfici covariates, true longitudinal trajectories and estimated longitudinal trajectories for each subject.}
       \item{true.first.deriv}{a (num.grid/2) x n array of the first derivatives of the true longitudinal trajectories, where each column is the first derivatives of the true longitudinal responses corresponding subject-specific log scaled ages in \code{new.pred.data} for each subject.}
       \item{true.second.deriv}{a (num.grid/2) x n array of the second derivatives of the true longitudinal trajectories, where each column is the second derivatives of the true longitudinal responses corresponding subject-specific log scaled ages in \code{new.pred.data} for each subject.}
       \item{est.first.deriv}{a (num.grid/2) x n array of the estiamted first derivatives of longitudinal trajectories, where each column is the estimated second derivatives of longitudinal responses corresponding subject-specific log scaled ages in \code{new.pred.data} for each subject.}
       \item{est.second.deriv}{a (num.grid/2) x n array of the estiamted second derivatives of longitudinal trajectories, where each column is the estimated second derivatives of longitudinal responses corresponding subject-specific log scaled ages in \code{new.pred.data} for each subject.}
       \item{boot.est.logT}{a n x num.boot array of the bootstrap estimated inflection points, where each row is a num.boot-length of boostrap estimates of the inflection point for each subject.}
       \item{ind.sd}{the n-length of the estimated boostrap standard deviations, where each element is the estimated standard deviation of the bootstrap estimates for each subject (each row of \code{boot.est.logT}).}
       \item{cp.boot}{a n x 2 array of the 95\% bootstrap confidence intervals, where each row has the lower bound and the upper bound of the 95\% confidence interval for the individual inflection point (i.e., 25th and 97.5th percentiles of the increasing ordered boostrap estimates for each row of \code{boot.est.logT}).}
}
\description{
Parametric nonlinear mixed effects model (NLME) approach: When true data are generated by the logistic model, the parametric NLME procedure is performed under the (missepcified) arctangent model .
}
\examples{

library(HDChangePoint)


## Specify parameters to generate true data
n=80;
model="logist";
p=2;
bb0=0.5;
bb=0.1;
x.sd=0.3;
v1=5;
v2=7;
dist="normal";
eps.sd=0.05;
u.sd=0.05;

## Specify parameters for the parametric NLME procedure

num.boot=1000;
true.theta1=6;
true.theta3=1;
num.grid=45;
eps.sd=0.05;
dist="normal";
para1=1;
para2=3.5;
para3=0;
para4=3;
para5=1;

## generate data with seed number

set.seed(1)
outdat<-mydata(n=n, model=model, p=p, bb0=bb0, bb=bb, x.sd=x.sd, v1=v1, v2=v2, dist=dist, eps.sd=eps.sd, u.sd=u.sd)

## Do parametric NLME estimation
results<-main.arctan.missp.nlme(n=n, model=model,  dat=outdat, num.boot=num.boot, num.grid=num.grid,
                                true.theta1=true.theta1, true.theta3=true.theta3, para1=para1, para2=para2,
                                para3=para3, para4=para4, para5=para5,  eps.sd=eps.sd, dist=dist)



}
