% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_results_summary_missp_logist.R
\name{missp.logist.results}
\alias{missp.logist.results}
\title{Summary results of the parametric NLME, when the true data are generated by the arctangent model, but the parametric NLME assumes the incorrect logistic model.}
\usage{
missp.logist.results(nsim = 100, n = 80, res = results,
  time.length = 20, beta = 0.1, beta0 = 2, file = "nlme_logisic")
}
\arguments{
\item{nsim}{number of simulation runs.}

\item{n}{number of sample size.}

\item{res}{a nsim-length of list for simulation results.}

\item{time.length}{number of data points at which predictors are required for each individual longitudinal trajectory. This time point for graphs to be plotted.}

\item{beta}{a (p-1)-length of true coefficient vector corresponding to subject specific covariates.}

\item{beta0}{true intercept of the log-normal model for the inflection point.}

\item{file}{a character string of file name for a creating plot in .eps file extention.}
}
\value{
A list of the summarized simulation results including

       \item{res.table}{a data frame of the absolute biases, estimated standard deviations, average of the estimated standard errors and 95\% coverage probabilities for the fixed effects (\code{beta0}, \code{beta}).}
       \item{Rel.glob.logT.est}{a data frame of the relative average performance of inflection points for all subjects including the relative absolute biases, the relative empirical standard errors, the relative bootstrap standard deviations and 95\% boostrap confidence intervals.}
}
\description{
Summary results of the parametric NLME, when the true data are generated by the arctangent model, but the parametric NLME assumes the incorrect logistic model.
}
\examples{

library(HDChangePoint)
data(missp_logist_visit1)

## Specify parameters to obtain the summary results

nsim=100;
n=80;
time.length=20;
beta=0.1;
beta0=2;
file="nlme_arctan_"

## summary results
simus.missp.logist.out<-missp.logist.results(nsim=nsim, n=n, res=results, time.length=time.length ,  beta=beta, beta0=beta0,  file=file)


}
