% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dat.R
\name{mydata}
\alias{mydata}
\title{Generate A Longitudinal Data Set}
\usage{
mydata(n = 80, model = "logist", p = 2, bb0 = 0.1, bb = 0.5,
  x.sd = 0.5, v1 = 5, v2 = 7, dist = "normal", eps.sd = 0.05,
  u.sd = 0.05)
}
\arguments{
\item{n}{number of sample size.}

\item{model}{a character string for a nonlinear model: \code{"logist"} or \code{"arctan"}.}

\item{p}{a parameter for the dimension of coefficients in the log-normal model for the inflection point.}

\item{bb0}{a parameter for the true intercept term of the log-normal model for the inflection point.}

\item{bb}{a (p-1)-length of the true coefficient vector of subject specific covariates.}

\item{x.sd}{a parameter for the standard deviation of time points (log-transformed ages).}

\item{v1}{a parameter for the smallest visit number among all subjects have made.}

\item{v2}{a parameter for the largest visit number among all subjects have made.}

\item{dist}{a character string for the distribution of the within-subject error term in the longitudinal model. Default is \code{"normal"}.}

\item{eps.sd}{a true scale parameter of the within-subject error term in the longitudinal model.}

\item{u.sd}{a true scale parameter of the random error term in the inflection point model.}
}
\value{
A data frame of the longitudinal data with 7 variables:

       \item{subj.id}{a vector of the ID number for all subjects.}
       \item{subj.cov}{ a vector of the subject specific covariates for all subjects.}
       \item{true_logt}{a vector of the inflection points for all subjects.}
       \item{ages}{a vector of the time points, which are the log transformed ages at clinical visits for all subjects}
       \item{omega}{a vector of the S-shaped function values depending on the ages and true_logt for all subjects.}
       \item{tms}{a vector of the observed longitudinal responses (total motor scores) for all subjects}
       \item{sex}{a vector of the subjects's gender.}

       The number of rows for the returned data frame is determined by the number of subjects' visits,
       where 5, 6, and 7 visits are randomly assigned to all patients. For example, the first subject made 5 visits and the second subject made 7 visits. Then the size of the data frame containing these two subjects' longitudinal data is 12 by 7.
}
\description{
Generate A Longitudinal Data Set
}
\examples{

library(HDChangePoint)

## Specify parameters to generate data

n=80;
model="logist";
p=2;
bb0=0.5;
bb=0.1;
x.sd=0.3;
v1=5;
v2=7;
dist="normal";
eps.sd=0.05;
u.sd=0.05;

## Generate a data set

set.seed(22)
outdata<-mydata(n=n, model=model, p=p, bb0=bb0, bb=bb, x.sd=x.sd, v1=v1, v2=v2, dist=dist, eps.sd=eps.sd, u.sd=u.sd)



}
