% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpara_derivative.R
\name{sec.der.mid}
\alias{sec.der.mid}
\title{Title}
\usage{
sec.der.mid(x, y)
}
\arguments{
\item{y}{}
}
\value{
Given vectors x and y, returns second derivatives of vector y corresponding a vector x
}
\description{
Title
}
\examples{
x<-seq(-0.5, 0.5, by=0.01); y<-w(x, 0, model="logist");
          out<-sec.der.mid(x, y); plot(out);


}
