% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stage_nonpara.R
\name{sim.nonpara}
\alias{sim.nonpara}
\title{Multi-Stage Nonparametric Procedure}
\usage{
sim.nonpara(n = 80, model = "arctan", dist = "normal", k1 = 10,
  k2 = 10, num.interp = 50, newl = 50, eps.sd = 0.05,
  mean.diff = 1, tolerance = 0.009, itermax = 20, iter = 0,
  time.length = 20, dat = outdat)
}
\arguments{
\item{n}{number of sample size.}

\item{model}{a character string for a nonlinear model: \code{"logist"} or \code{"arctan"}.}

\item{dist}{a character string for the distribution of within-subject error term in the longitudinal model. Default is \code{"normal"}.}

\item{k1}{a parameter for the dimension of the basis functions in order to represent the monotone increasing and convex smooth function.}

\item{k2}{a parameter for the dimension of the basis functions in order to represent the monotone increasing and concave smooth function.}

\item{num.interp}{number of pseudo-data points, which are unifromly generated within the observed data points.}

\item{newl}{a length of time points (log transformed ages) at which predictors are required for each individual longitudinal trajectory.}

\item{eps.sd}{a true scale parameter of the within-subject error term in the longitudinal model.}

\item{mean.diff}{an initial tolerence for the convergence criterion. Default is 1.}

\item{tolerance}{a parameter for the tolerence for the convergence criterion. Default is 0.009.}

\item{itermax}{a parameter for the maximum number of iterations in the multi-stage nonparametric algorithm. Default is 50.}

\item{iter}{initial number of iteration in the multi-stage nonparametric algorithm. The default is 0.}

\item{time.length}{a length of time points for each individual longitudinal trajectory graph to be plotted, should be less than \code{newl}.}

\item{dat}{a data frame of the generated data set.}
}
\value{
A list of the simulation results includes
       \item{true_z}{a n-length of the true individual inflection points vector.}
       \item{zstar0}{a n-length of the estimated initial individual inflection points vector.}
       \item{zstar}{a n-length of the estimated individual inflection points vector.}
       \item{indvidual.Ta}{a n-length of estimated individual inflection points vector after all trajectories were shifted by their inflection points.}
       \item{newlogS}{a newl-length of the new time points vector at which predictors are required for all subjects. }
       \item{tms.pred}{a newl-length of the predcited longitudinal trajectory corresponding \code{newlogS} from all subjects' longitudinal data. }
       \item{tms.se.pred}{a newl-length of the estimated standard errors for \code{tms.pred}.}
       \item{global.T}{estimated inflection point on the shifted time points for all subjects.}
       \item{est.sigma.epsi}{an estimated scale parameter of the within-subject error in the longitudinal model.}
       \item{ind.b.sd}{a n-length of the boostrap standard deviations vector of \code{zstar}, where each component is the bootstrap standard deviaion of the estimated inflection point for each subject. }
       \item{ind.est.cp}{A n x 2 95\% bootstrap confidence intervals matrix for \code{zstar}, i.e., each row has the lower bound and the upper bound of the 95\% bootstrap confidence intervals for each individual's estimated inflection point. }
       \item{est.beta0}{estimated intercept parameter of the log-normal model for the inflection point.}
       \item{str.beta0}{estimated standard error of \code{est.beta0}.}
       \item{est.beta}{a (p-1)-length of estimated coefficient vector of subject specific covariates in the log-normal model for the inflection point.}
       \item{str.beta}{a (p-1)-length of vector with the estimated standard errors of \code{est.beta}. }
       \item{est.sigma.ei}{an estimated scale parameter of the random error in the log-normal model for the inflection point. }
       \item{xx.time}{A time.length x n time points matrix, where each column is a subset of the each column of \code{newdlogS}, i.e., the individual time points at which estimated numerical derivatives are required for each subject.}
       \item{ind.org.tms}{A time.length x n true longitudinal trajectories matrix, where  each column is the true individual longitudinal trajectory at each column of \code{xx.time} for each subject.}
       \item{ind.est.tms}{A time.length x n estimated longitudinal trajectories matrix, where  each column is the subset of \code{ind.predict.tms}, i.e., estimated individual longitudinal trajectory at each column of \code{xx.time} for each subject. }
       \item{trueD1}{A time.length x n numerically estimated first derivatives matrix of \code{ind.org.tms}, where each column is the numerically estimated first derivative of each column of \code{ind.org.tms} at each column of \code{xx.time} for each subject. }
       \item{trueD2}{A time.length x n numerically estimated second derivatives matrix of \code{ind.org.tms}, where each column is the numerically estimated second derivative of each column of \code{ind.org.tms} at each column of \code{xx.time} for each subject. }
       \item{predD1}{A time.length x n numerically estimated first derivatives matrix of \code{ind.est.tms}, where each column is the numerically estimated first derivative of each column of \code{ind.est.tms} at each column of \code{xx.time} for each subject.}
       \item{predD2}{A time.length x n numerically estimated second derivatives matrix of \code{ind.est.tms}, where each column is the numerically estimated second derivative of each column of \code{ind.est.tms} at each column of \code{xx.time} for each subject.}
       \item{iter}{number of interations until convergence. }
       \item{mean.diff}{estimated convergence criterion, which is less than a \code{tolerance} if the algorithm convergences.}
}
\description{
Multi-Stage Nonparametric Procedure
}
\examples{

library(HDChangePoint)

## Specify parameters to generate true data
n=80;
model="logist";
p=2;
bb0=0.5;
bb=0.1;
x.sd=0.3;
v1=8;
v2=10;
dist="normal";
eps.sd=0.05;
u.sd=0.05;

## Specify parameters for the multi-stage nonparametric procedure
num.interp=45;
newl=45;
k1=20;
k2=20;
tolerance=0.009
iter=0;
time.length=20;

set.seed(22)

## Data generation under the logistic model
outdat<-mydata(n=n, model=model, p=p, bb0=bb0, bb=bb, x.sd=x.sd,  dist=dist,v1=v1, v2=v2, eps.sd=eps.sd, u.sd=u.sd) #


## Multi-stage nonparametric estimation
results<-sim.nonpara(n=n, model=model, dist=dist, k1=k1, k2=k2, num.interp=num.interp, newl=newl, 
                      eps.sd=eps.sd, mean.diff=1, tolerance=tolerance,
                     itermax=50, iter=iter, time.length=time.length, dat=outdat)


}
