% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stage_nonpara.R
\name{simus.results}
\alias{simus.results}
\title{Title}
\usage{
simus.results(nsim = 100, n = 80, res = results, model = "logist",
  num.interp = 45, newl = 45, bb.true = 0.1, b0.true = 0.5,
  true.sigma.u = 0.05, true.sigma.eps = 0.05, num.interval = 18,
  file = "logist")
}
\arguments{
\item{file}{}
}
\value{
summarize simulation results from the multi-stage nonparametric procedure including the performance of longituidinal model parameters
        , inflection points (absolute biases, estiamted standard deviations, average of the estimated standard errors and 95% coverage probabilities)
        and plots for the average trends of estimated longitudinal trajectories and their sencond derivatives.
}
\description{
Title
}
\examples{
 simus.out<-simus.results(nsim=100, n=80, res=results, model="logist", num.interp=45, newl=45, bb.true=0.1,
                                  b0.true=0.5, true.sigma.u=0.05, true.sigma.eps=0.05, num.interval=18, file="logist")


}
