% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_omega_ft.R
\name{w}
\alias{w}
\title{Nonlinear S-shaped Model}
\usage{
w(x, logT, model)
}
\arguments{
\item{x}{a vector of time points. It can be any length of a vector.}

\item{logT}{a parameter for an inflection point.}

\item{model}{a character string for a nonlinear model: \code{"logist"} or \code{"arctan"}.}
}
\value{
a vector of response values corresponding a vector of time points \eqn{x}, whose trajecotry follows an S-shape.
        Two deterministic \code{"S"}\code{-}shaped functions are the logstic and the aractangent functions and user can choose either one.
}
\description{
Nonlinear S-shaped Model
}
\examples{

library(HDChangePoint)


## Generate a vector of time points and an inflection point in the domain of the time points
x<-seq(-0.5, 0.5, by=0.1); logT<-0;

## Obtain an S-shaped function under the logistic model
out<-w(x, logT, model="logist");
plot(out)

## Obtain an S-shaped function under the arctangent model
out1<-w(x, logT, model="arctan");
plot(out1)


}
